change sql

+++ wx-erp-springboot/src/main/kotlin/io/github/mellivorines/wxerp/repository/MetaRepository.kt
@@ -1,20 +1,19 @@
 package io.github.mellivorines.wxerp.repository
 
-import io.github.mellivorines.wxerp.entity.Category
 import io.github.mellivorines.wxerp.entity.Meta
 import org.babyfish.jimmer.spring.repository.KRepository
 import org.springframework.stereotype.Repository
 
 /**
  * <p>
- * CategoryRepository 接口
+ * MetaRepository 接口
  * </p>
  *
- * @author lilinxi
- * @date 2023-04-11
+ * @author mellivorines
+ * @date 2023-04-24
  */
 @Repository
-interface MetaRepository : KRepository<Meta, Long> {
+interface MetaRepository : KRepository<Meta, Int> {
 
 }
